rest-api/requirements.txt
waitress==1.3.1
Django==2.2.6
drf_nested_routers==0.91
django_allauth==0.40.0
django_rest_auth==0.9.5
protobuf==3.10.0
PyJWT==1.7.1
django-cors-headers
psycopg2
pillow
djangorestframework-jwt
gunicorn==19.3.0
django-cors-middleware